+++ src/components/PageFooter.tsx
@@ -12,11 +12,10 @@ export const PageFooter = () => (
         <Text>此版本为<Text code>{VERSION}</Text>正式版</Text>
       </div>
       <div>
-        <span>注：源码展示出来仅供参考</span>
-        <Link
+        <span>注：<Link
           target="_blank"
           href="https://github.com/shinnku-nikaidou/upset-gal-web"
-        />
+        > 源码 </Link>展示出来仅供参考</span>
       </div>
     </Text>
   </Footer>
